requirements.txt
Django==3.1.4
djangorestframework==3.12.2
psycopg2-binary==2.9.7
redis==3.5.0
PyJWT==2.8.0